+++ tests/api/runner/request_executor.rb
@@ -1,0 +1,16 @@
+class RequestExecutor
+  def initialize(request)
+    @request = request
+  end
+
+  def call
+    uri = "http://localhost:5000#{request[:uri]}"
+
+    HTTP.send(request[:verb].downcase, uri)
+  end
+
+  private
+
+  attr_reader :request
+end
+
